fix: ColorString split to Printer & UserInterface User input colour name instead of color number
